+++ src/renderer/program.h
@@ -14,16 +14,16 @@ using namespace std;
 
 namespace std {
 	template<>
-	struct hash<pair<string, size_t>> {
-		size_t operator()(pair<string, size_t> const& source) const noexcept {
-			size_t result = hash<string>{}(source.first);
+	struct hash<pair<string, uint64_t>> {
+		size_t operator()(pair<string, uint64_t> const& source) const noexcept {
+			uint64_t result = hash<string>{}(source.first);
 			return result ^ (source.second + 0x9e3779b9 + (result << 6) + (result >> 2));
     }
 	};
 
 	template<>
-	struct equal_to<pair<string, size_t>> {
-		bool operator()(const pair<string, size_t>& x, const pair<string, size_t>& y) const {
+	struct equal_to<pair<string, uint64_t>> {
+		bool operator()(const pair<string, uint64_t>& x, const pair<string, uint64_t>& y) const {
 			return x.first == y.first && x.second == y.second;
 		}
 	};
@@ -35,7 +35,7 @@ namespace render {
 			Program(class Window* window);
 			void bind();
 			void addShader(class Shader* shader);
-			void bindUniform(string uniformName, void* data, unsigned int size, size_t cacheIndex = 0, bool setOnce = false);
+			void bindUniform(string uniformName, void* data, unsigned int size, uint64_t cacheIndex = 0, bool setOnce = false);
 			void bindTexture(string uniformName, unsigned int texture);
 		
 		protected:
@@ -49,9 +49,9 @@ namespace render {
 			void createUniformMemory(string uniformName, unsigned int size);
 			#else
 			GLuint program = GL_INVALID_INDEX;
-			tsl::robin_map< pair<string, size_t>, GLuint> uniformToBuffer;
+			tsl::robin_map< pair<string, uint64_t>, GLuint> uniformToBuffer;
 			static unsigned int UniformCount;
-			void createUniformBuffer(string uniformName, unsigned int size, size_t cacheIndex);
+			void createUniformBuffer(string uniformName, unsigned int size, uint64_t cacheIndex);
 			#endif
 	};
 };
