+++ src/resources/resourceManager.h
@@ -16,28 +16,28 @@ namespace es {
 	esEntryPtr hotReload(esEnginePtr esEngine, unsigned int argc, esEntryPtr args);
 };
 
-void handleSpritesheets(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handlePNGs(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handleHTML(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handleCSS(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handleScripts(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handleShaders(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handleDKSHShaders(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handleMaps(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-void handleSounds(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
+void handleSpritesheets(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handlePNGs(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handleHTML(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handleCSS(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handleScripts(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handleShaders(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handleDKSHShaders(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handleMaps(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+void handleSounds(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
 
 namespace resources {
 	class ResourceManager {
 		friend ResourceObject;
-		friend void ::handleSpritesheets(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handlePNGs(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handleHTML(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handleCSS(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handleScripts(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handleShaders(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handleDKSHShaders(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handleMaps(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
-		friend void ::handleSounds(void* owner, carton::File* file, const char* buffer, size_t bufferSize);
+		friend void ::handleSpritesheets(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handlePNGs(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handleHTML(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handleCSS(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handleScripts(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handleShaders(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handleDKSHShaders(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handleMaps(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
+		friend void ::handleSounds(void* owner, carton::File* file, const char* buffer, uint64_t bufferSize);
 		
 		public:
 			ResourceManager(string fileName);
