+++ src/resources/shaderSource.h
@@ -5,13 +5,13 @@
 namespace resources {
 	class ShaderSource : public ResourceObject {
 		public:
-			ShaderSource(class ResourceManager* manager, carton::Metadata* metadata, const unsigned char* buffer, size_t bufferSize, bool useCharBuffer =  false);
+			ShaderSource(class ResourceManager* manager, carton::Metadata* metadata, const unsigned char* buffer, uint64_t bufferSize, bool useCharBuffer =  false);
 			string source;
 			unsigned char* buffer;
-			size_t bufferSize = 0;
+			uint64_t bufferSize = 0;
 			ShaderSource* original = nullptr;
 
-			void reload(carton::Metadata* metadata, const unsigned char* buffer, size_t bufferSize) {
+			void reload(carton::Metadata* metadata, const unsigned char* buffer, uint64_t bufferSize) {
 				ResourceObject::reload(metadata, buffer, bufferSize);
 				printf("shader source reload not implemented\n");
 			}
